Commented out the old browser warning stuff.
The solution that we use sucks because when you try to share our pages on
Facebook, it qualifies as an old browser and picks up this text in the website
preview that is auto-generated by Facebook via scraping.

We would need to add a long list of bot-user-agents that will never see this
warning and I don't know how we could possibly keep this list up-to-date.